flask_server/requirements.txt
flask
joblib
nltk==3.8.1
scikit-learn
scipy
transformers
langdetect
sentencepiece
torch
PyPDF2
python-docx
# underthesea
pymupdf
torchvision
mysql-connector-python